Rework detection of multipath disks to remove in the installer
Checking whether "somestring" in "somestring" leads to bad
results when we check if "sda" is part of "sdaa sdab...". Don't
spuriously remove valid disks. Use a dict instead and check
against that. Log the original output, just in case it's ever
needed anywhere.

Also, use a set to ensure unique values instead of looping through
and decrementing the count.

Change-Id: I20be68e232e25e9a751516edbaece69c0b81107d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1203176
Signed-off-by: Ryan Barry <rbarry@redhat.com>